+++ lib/utils/models/selectable_word_item.dart
@@ -1,0 +1,5 @@
+class SelectableWordItem {
+  int id;
+  String keyword;
+  SelectableWordItem(this.id, this.keyword);
+}
